+++ src/core/wi_api.c
@@ -268,43 +268,43 @@ wi_pop_userdata(struct wi_state* state, const char* name) {
 }
 
 bool
-wi_arg_is_real(struct wi_state* state, int arg) {
+wi_arg_is_real(struct wi_state* state, uint8_t arg) {
     return wi_value_is_real(state->ffi_stack[arg]);
 }
 
 bool
-wi_arg_is_null(struct wi_state* state, int arg) {
+wi_arg_is_null(struct wi_state* state, uint8_t arg) {
     return wi_value_is_null(state->ffi_stack[arg]);
 }
 
 bool
-wi_arg_is_bool(struct wi_state* state, int arg) {
+wi_arg_is_bool(struct wi_state* state, uint8_t arg) {
     return wi_value_is_bool(state->ffi_stack[arg]);
 }
 
 bool
-wi_arg_is_string(struct wi_state* state, int arg) {
+wi_arg_is_string(struct wi_state* state, uint8_t arg) {
     return wi_value_is_string(state->ffi_stack[arg]);
 }
 
 bool
-wi_arg_is_function(struct wi_state* state, int arg) {
+wi_arg_is_function(struct wi_state* state, uint8_t arg) {
     wi_value value = state->ffi_stack[arg];
     return wi_value_is_foreign(value) || wi_value_is_closure(value);
 }
 
 bool
-wi_arg_is_object(struct wi_state* state, int arg) {
+wi_arg_is_object(struct wi_state* state, uint8_t arg) {
     return wi_value_is_object(state->ffi_stack[arg]);
 }
 
 bool
-wi_arg_is_userdata(struct wi_state* state, int arg, const char* name) {
+wi_arg_is_userdata(struct wi_state* state, uint8_t arg, const char* name) {
     return _is_userdata(state->ffi_stack[arg], name);
 }
 
 wi_real
-wi_arg_real(struct wi_state* state, int arg) {
+wi_arg_real(struct wi_state* state, uint8_t arg) {
     if (WI_UNLIKELY(!wi_arg_is_real(state, arg))) {
         wi_state_error(state, "bad argument %i - expected a value of type real but got %s", arg,
                        wi_value_type(state->ffi_stack[arg]));
@@ -314,7 +314,7 @@ wi_arg_real(struct wi_state* state, int arg) {
 }
 
 void
-wi_arg_null(struct wi_state* state, int arg) {
+wi_arg_null(struct wi_state* state, uint8_t arg) {
     if (WI_UNLIKELY(!wi_arg_is_null(state, arg))) {
         wi_state_error(state, "bad argument %i - expected a value of type null but got %s", arg,
                        wi_value_type(state->ffi_stack[arg]));
@@ -322,7 +322,7 @@ wi_arg_null(struct wi_state* state, int arg) {
 }
 
 bool
-wi_arg_bool(struct wi_state* state, int arg) {
+wi_arg_bool(struct wi_state* state, uint8_t arg) {
     if (WI_UNLIKELY(!wi_arg_is_bool(state, arg))) {
         wi_state_error(state, "bad argument %i - expected a value of type bool but got %s", arg,
                        wi_value_type(state->ffi_stack[arg]));
@@ -332,7 +332,7 @@ wi_arg_bool(struct wi_state* state, int arg) {
 }
 
 char*
-wi_arg_string(struct wi_state* state, int arg, int* count, int* len) {
+wi_arg_string(struct wi_state* state, uint8_t arg, int* count, int* len) {
     if (WI_UNLIKELY(!wi_arg_is_string(state, arg))) {
         wi_state_error(state, "bad argument %i - expected a value of type string but got %s", arg,
                        wi_value_type(state->ffi_stack[arg]));
@@ -352,7 +352,7 @@ wi_arg_string(struct wi_state* state, int arg, int* count, int* len) {
 }
 
 void
-wi_arg_function(struct wi_state* state, int arg, uint8_t arity) {
+wi_arg_function(struct wi_state* state, uint8_t arg, uint8_t arity) {
     wi_value function = state->ffi_stack[arg];
 
     if (!wi_value_is_foreign(function) && !wi_value_is_closure(function)) {
@@ -372,7 +372,7 @@ wi_arg_function(struct wi_state* state, int arg, uint8_t arity) {
 }
 
 struct wi_object*
-wi_arg_object(struct wi_state* state, int arg) {
+wi_arg_object(struct wi_state* state, uint8_t arg) {
     wi_value value = state->ffi_stack[arg];
 
     if (WI_UNLIKELY(!wi_value_is_object(value))) {
@@ -384,7 +384,7 @@ wi_arg_object(struct wi_state* state, int arg) {
 }
 
 void*
-wi_arg_userdata(struct wi_state* state, int arg, const char* name) {
+wi_arg_userdata(struct wi_state* state, uint8_t arg, const char* name) {
     wi_value value = state->ffi_stack[arg];
 
     if (WI_UNLIKELY(!wi_arg_is_userdata(state, arg, name))) {
